+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 =========
-
+### v0.0.2, 27 Dec 2020, Elixir `~> 1.11`
+* Added `Dumper` Module to communicate with linux can sockets
+* Added `Mock` Module for simulating can messages
+* Added `Tools` Module to rewrap existing functions that were previously in the `Cannes` Module
 ### v0.0.1, 25 Dec 2020, Elixir `~> 1.11`
 * Alpha Version release for Cannes.